+++ CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-## Changed
+## [0.10.26] - 2025-05-16
 
-- ODCS export: Export physical type if the physical type is configured in config object
+### Changed
+
+- Databricks: Add support for Variant type (#758)
+- `datacontract export --format odcs`: Export physical type if the physical type is configured in
+  config object (#757)
+- `datacontract export --format sql` Include datacontract descriptions in the Snowflake sql export (
+  #756)
 
 ## [0.10.25] - 2025-05-07